+++ client/src/pages/common/LoginPage.js
@@ -1,7 +1,7 @@
 import React, { useState } from "react"
 import { motion } from "framer-motion"
 import { Lock, Mail, Eye, EyeOff } from 'lucide-react'
-import { useAuth } from '../context/AuthContext'
+import { useAuth } from '../../context/AuthContext'
 import { useNavigate, Link } from 'react-router-dom'
 
 const Login = () => {
