+++ package.json
@@ -27,7 +27,7 @@
     "google-auth-library": "^9.15.0",
     "jsonwebtoken": "^9.0.2",
     "lucide-react": "^0.464.0",
-    "mongoose": "^8.2.2",
+    "mongoose": "^8.8.4",
     "multer": "^1.4.5-lts.1",
     "react-chartjs-2": "^5.2.0"
   },
